// src/models/file-properties-conseil.ts
export interface FilePropertiesConseil {
  aTraiter: string;
  codeBudget: string;
  codePostal: string;
  dateContrat: string;
  dateDocument: string;
  dateFinContrat: string;
  dateReception: string;
  dateSignature: string;
  description: string;
  documentType: string;
  famille: string;
  fournisseur: string;
  matriculeCollab: string;
  nature: string;
  nomClient: string;
  numeroClient: string;
  numeroContrat: string;
  numeroDossier: string;
  produit: string;
  sousFamille: string;
  SCPI?: string;
}