// src/enums/groups.enums.ts
export enum GroupsEnums {
  isUserMR = 'GROUP_GRP_EXPERTISE',
  isUserConseilBO = 'GROUP_GRP_CONSEIL_BO',
  isUserConseilDRSiege = 'GROUP_GRP_CONSEIL_DR_SIEGE',
  isUserConseilCGP = 'GROUP_GRP_CONSEIL_CGP',
  isUserConseilDRPlateforme = 'GROUP_GRP_CONSEIL_DR_PLATEFORME',
  isUserADV = 'GROUP_GRP_ADV',
  isUserGeranceAssociesPartenaire = 'GROUP_GRP_GERANCE_ASSOCPART',
  isUserGeranceAssociesAssociesBO = 'GROUP_GRP_GERANCE_ASSOC_BO',
  isUserGeranceAssociesAssociesSUPP = 'GROUP_GRP_GERANCE_ASSOC_SUPP',
  isUserGeranceAssociesAssociesCONSULT = 'GROUP_GRP_GERANCE_ASSOC_CONSULT',
  isUserGeranceAssociesAssociesCONSULT_RECL = 'GROUP_GRP_GERANCE_ASSOC_CONSULT_RECL',
  isUserGeranceAssociesAssociesRECL = 'GROUP_GRP_GERANCE_ASSOC_RECL',
  isUserGeranceAssociesPartenaireBO = 'GROUP_GRP_GERANCE_PART_BO',
  isUserGeranceAssociesPartenaireSUPP = 'GROUP_GRP_GERANCE_PART_SUPP',
  isUserSofiral = 'GROUP_GRP_SOFIRAL',
  isUserArchivesPresidenceARCHIVISTES = 'GROUP_GRP_ARCHIVESPRESIDENCE_ARCHIVISTES',
  isUserArchivesPresidenceCONSULT = 'GROUP_GRP_ARCHIVESPRESIDENCE_CONSULTATION',
  isUserFidusignAchat = 'GROUP_GRP_FSG_ACHAT',
  isUserFidusignAchatAccesSpec = 'GROUP_GRP_FSG_ACHAT_ACCES_SPEC',
  isUserGeranceImmobilier = 'GROUP_GRP_GERANCE_IMMOBILIER',
  isUserBudgetPlusGestion = 'GROUP_GRP_BUDGETPLUS_GESTION',
  isUserGeranceAssociesAssociesCONTROL = 'GROUP_GRP_GERANCE_ASSOC_CONTROLE',
  isUserFidusignJuridique = 'GROUP_GRP_FSG_JURIDIQUE_FIRME',
  isUserFidusignJuridiqueConsult = 'GROUP_GRP_FSG_JURIDIQUE_FIRME_CONSULT',
  isUserFidusignYProximite = 'GROUP_GRP_FSG_YPROXIMITE',
  isUserFidusignYProximiteCommercial = 'GROUP_GRP_FSG_YPROXIMITE_COMMERCIAL',
  isUserFidusignYProximiteConsult = 'GROUP_GRP_FSG_YPROXIMITE_CONSULT',
  isUserExportDap = 'GRP_RH_EXPORT_PERTEMARCHE_DAP',
  isUserExportFormation = 'GRP_RH_EXPORT_PERTEMARCHE_FORMATION',
  isUserExportBulletins = 'GRP_RH_EXPORT_BULLETINS',
  isUserExportAutres = 'GRP_RH_EXPORT_AUTRES_DOCUMENTS',
}